requirements.txt
altgraph==0.17.4
cffi==1.17.1
cryptography==44.0.2
gitdb==4.0.12
GitPython==3.1.44
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.1
pyuic5-tool==0.0.1
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
tzdata==2025.1
